demo/requirements.txt
django-bootstrap-static==2.3.1
feedparser==5.1.3
hadrian==1.1.5
lxml
django-registration
django-social-auth==0.7.22
